Cell groundtruth

A matrix with the cell type proportions of the single cell object.

Usage

cells_groundtruth

Format

Matrix with samples as rows and cell types as columns

Examples

data(cells_groundtruth)
head(cells_groundtruth)
#>                 B.cells CD4.cells CD8.cells Endothelial      Cancer Fibroblasts
#> Sample_11522 0.12895881 0.3452701 0.3183606 0.010970813 0.002276961 0.007037880
#> Sample_11817 0.10644728 0.1799866 0.3425118 0.003022163 0.016453996 0.002686367
#> Sample_11918 0.17508418 0.3804714 0.2188552 0.000000000 0.006734007 0.000000000
#> Sample_12889 0.19860627 0.2195122 0.1522648 0.001742160 0.008710801 0.031707317
#> Sample_12929 0.06085377 0.2933697 0.1889192 0.006357856 0.013623978 0.079927339
#> Sample_12935 0.11433757 0.2934059 0.2692075 0.003629764 0.010889292 0.007259528
#>                Mast.cells Mural.cells Myeloid.cells    NK.cells     Plasma
#> Sample_11522 0.0629269302 0.003725937    0.02835852 0.005588905 0.02173463
#> Sample_11817 0.0040295500 0.005372733    0.06447280 0.024848892 0.04163868
#> Sample_11918 0.0000000000 0.000000000    0.13804714 0.013468013 0.00000000
#> Sample_12889 0.0003484321 0.011846690    0.01463415 0.008362369 0.19059233
#> Sample_12929 0.0027247956 0.025431426    0.11807448 0.019073569 0.06357856
#> Sample_12935 0.0006049607 0.006654567    0.06594071 0.019963702 0.04053237
#>              T.cells.proliferative T.cells.helper T.cells.regulatory
#> Sample_11522           0.006002898     0.02090664         0.03788036
#> Sample_11817           0.048018805     0.08226998         0.07824043
#> Sample_11918           0.013468013     0.02693603         0.02693603
#> Sample_12889           0.012891986     0.07351916         0.07526132
#> Sample_12929           0.004541326     0.02906449         0.09445958
#> Sample_12935           0.015124017     0.06291591         0.08953418