Pseudobulk

A gene expression matrix after aggregation of the single cell object (pseudobulk).

Usage

pseudobulk

Format

Matrix with genes as rows and samples as columns

Examples

data(pseudobulk)
head(pseudobulk)
#>           Sample_12929 Sample_15467 Sample_13634 Sample_14428 Sample_11817
#> FAM87B        0.235074    0.1148403   0.05574872    0.0000000    0.0000000
#> LINC00115     7.362652    6.8824045   8.61430844    5.4616572    5.7012964
#> FAM41C        0.000000    0.1460498   0.00000000    0.3789071    0.6343796
#> SAMD11        3.817682   13.1919087   2.54044243   20.2266121    0.4972594
#> NOC2L        49.262883   42.5503868  45.21650821   40.6154460   43.4948009
#> KLHL17        6.951867    2.2172485   3.52455121    2.7615027    5.7434283
#>           Sample_14958 Sample_12935 Sample_13636 Sample_13774 Sample_14965
#> FAM87B       0.3236136    0.0000000    0.0000000     0.000000    0.1560621
#> LINC00115   19.8165892    3.3543683   11.4059411     5.454424    5.9069536
#> FAM41C       0.0000000    0.0000000    0.9141174     0.000000    0.0000000
#> SAMD11      10.4658028    0.1366895    4.3368385     4.522387    0.9483388
#> NOC2L       38.5585358   48.6747091   58.4742593    38.114552   40.7385217
#> KLHL17       3.8164825    5.0258448    9.1296835     3.784092    5.4281847
#>           Sample_12889 Sample_8356 Sample_11918 Sample_11522 Sample_15002
#> FAM87B       0.0000000    0.000000     1.235270    0.1189222   0.04609012
#> LINC00115   12.8818526    6.313088     1.398791    6.7683707  11.48885574
#> FAM41C       0.2434332    0.000000     0.000000    0.5163624   0.00000000
#> SAMD11       6.9320731    0.000000     0.000000    0.5507367   1.46185908
#> NOC2L       52.4100099   57.772738    74.878058   66.1555013  44.07698679
#> KLHL17       4.5332535    6.860315    39.540936    3.0568978   4.06263336