Compute cell type processing

Compute cell type processing

Usage

compute.deconvolution.analysis(
  deconvolution,
  corr = 0.7,
  seed = NULL,
  cells_extra = NULL,
  file_name = NULL,
  return = FALSE
)

Arguments

deconvolution

Deconvolution output of compute.deconvolution() with features as columns and samples as rows

corr

Minimum correlation threshold for subgroupping the deconvolution features

seed

A numeric value to specificy the seed. This ensures reproducibility during the choice step of high correlated features.

cells_extra

A string specifying the cells names to consider and that are not including in the nomenclature of multideconv (see Readme)

file_name

A string specifying the file name of the .csv file with the deconvolution subgroups

return

Boolean value to whether return and saved the plot and csv files of deconvolution generated during the run inside the Results/ directory.

Value

A list containing

A matrix with the deconvolution after processing

The deconvolution subgroups per cell type

The deconvolution subgroups composition

The deconvolution groups discarded caused they are all belonging to the same method

The discarded features because they contain a high number of zeros across samples (> 90%)

Discarded features due to low variance across samples

Discarded cell types because they are not supported in the pipeline

High correlated deconvolution pairs (>high_corr)

Examples

data("deconvolution")

processed_deconvolution = compute.deconvolution.analysis(deconvolution, corr = 0.7, seed = 123)
#> Removing features with high zero number 90%...............................................................
#> 
#> Checking distribution of variances...............................................................
#> 
#> Chosen threshold is: 0.0002698755 
#> 
#> Saving variance distribution plot in Results/ folder
#> 
#> Removing 157 features with variance across samples below this threshold...............................................................
#> 
#> Splitting deconvolution features per cell type...............................................................
#> 
#> Finding group of features with high correlation between each other...............................................................
#> 
#> [1] "Highly correlated features (r>0.9): Quantiseq_B.cells, DeconRNASeq_CBSX.HNSCC.scRNAseq_B.cells, DWLS_CBSX.HNSCC.scRNAseq_B.cells, DeconRNASeq_CBSX.Melanoma.scRNAseq_B.cells, Epidish_CBSX.Melanoma.scRNAseq_B.cells, CBSX_CBSX.Melanoma.scRNAseq_B.cells, DWLS_CBSX.Melanoma.scRNAseq_B.cells, DeconRNASeq_CCLE.TIL10_B.cells, Epidish_CCLE.TIL10_B.cells, CBSX_CCLE.TIL10_B.cells, DWLS_CCLE.TIL10_B.cells, DeconRNASeq_TIL10_B.cells, Epidish_TIL10_B.cells, CBSX_TIL10_B.cells, DWLS_TIL10_B.cells"
#> Keeping only feature: DWLS_TIL10_B.cells
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_BPRNACan_B.cells, Epidish_BPRNACan_B.cells, CBSX_BPRNACan_B.cells, DWLS_BPRNACan_B.cells, DeconRNASeq_BPRNACan3DProMet_B.cells, Epidish_BPRNACan3DProMet_B.cells, CBSX_BPRNACan3DProMet_B.cells, DWLS_BPRNACan3DProMet_B.cells, DeconRNASeq_BPRNACanProMet_B.cells, Epidish_BPRNACanProMet_B.cells, CBSX_BPRNACanProMet_B.cells, DWLS_BPRNACanProMet_B.cells, DeconRNASeq_BSeqSC.Vanderbilt.scRNAseq_B.cells, Epidish_BSeqSC.Vanderbilt.scRNAseq_B.cells, DWLS_BSeqSC.Vanderbilt.scRNAseq_B.cells, Epidish_DWLS.Vanderbilt.scRNAseq_B.cells, CBSX_DWLS.Vanderbilt.scRNAseq_B.cells, DWLS_DWLS.Vanderbilt.scRNAseq_B.cells, DeconRNASeq_MOMF.Vanderbilt.scRNAseq_B.cells, DWLS_MOMF.Vanderbilt.scRNAseq_B.cells, BayesPrism_Vanderbilt_B.cells, Bisque_Vanderbilt_B.cells, MuSic_Vanderbilt_B.cells, SCDC_Vanderbilt_B.cells"
#> Keeping only feature: DWLS_BSeqSC.Vanderbilt.scRNAseq_B.cells
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_CBSX.HNSCC.scRNAseq_B.cells, CBSX_CBSX.HNSCC.scRNAseq_B.cells"
#> Keeping only feature: Epidish_CBSX.HNSCC.scRNAseq_B.cells
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_CBSX.NSCLC.PBMCs.scRNAseq_B.cells, Epidish_CBSX.NSCLC.PBMCs.scRNAseq_B.cells, CBSX_CBSX.NSCLC.PBMCs.scRNAseq_B.cells, DWLS_CBSX.NSCLC.PBMCs.scRNAseq_B.cells"
#> Keeping only feature: CBSX_CBSX.NSCLC.PBMCs.scRNAseq_B.cells
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_CBSX.Vanderbilt.scRNAseq_B.cells, Epidish_CBSX.Vanderbilt.scRNAseq_B.cells, CBSX_CBSX.Vanderbilt.scRNAseq_B.cells"
#> Keeping only feature: CBSX_CBSX.Vanderbilt.scRNAseq_B.cells
#> 
#> [1] "Highly correlated features (r>0.9): DWLS_CBSX.Vanderbilt.scRNAseq_B.cells, DeconRNASeq_DWLS.Vanderbilt.scRNAseq_B.cells"
#> Keeping only feature: DWLS_CBSX.Vanderbilt.scRNAseq_B.cells
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_LM22_B.memory.cells, Epidish_LM22_B.memory.cells, CBSX_LM22_B.memory.cells, DWLS_LM22_B.memory.cells"
#> Keeping only feature: CBSX_LM22_B.memory.cells
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_CBSX.HNSCC.scRNAseq_Macrophages.cells, Epidish_CBSX.HNSCC.scRNAseq_Macrophages.cells, CBSX_CBSX.HNSCC.scRNAseq_Macrophages.cells, DWLS_CBSX.HNSCC.scRNAseq_Macrophages.cells, Epidish_CBSX.Melanoma.scRNAseq_Macrophages.cells, CBSX_CBSX.Melanoma.scRNAseq_Macrophages.cells, DWLS_CBSX.Melanoma.scRNAseq_Macrophages.cells"
#> Keeping only feature: DWLS_CBSX.Melanoma.scRNAseq_Macrophages.cells
#> 
#> [1] "Highly correlated features (r>0.9): Quantiseq_Macrophages.M1, DeconRNASeq_CCLE.TIL10_Macrophages.M1, Epidish_CCLE.TIL10_Macrophages.M1, DWLS_CCLE.TIL10_Macrophages.M1, DeconRNASeq_TIL10_Macrophages.M1, Epidish_TIL10_Macrophages.M1, DWLS_TIL10_Macrophages.M1"
#> Keeping only feature: DWLS_TIL10_Macrophages.M1
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_BPRNACan_Macrophages.M2, DeconRNASeq_BPRNACan3DProMet_Macrophages.M2, DeconRNASeq_BPRNACanProMet_Macrophages.M2"
#> Keeping only feature: DeconRNASeq_BPRNACanProMet_Macrophages.M2
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_CCLE.TIL10_Macrophages.M2, DWLS_CCLE.TIL10_Macrophages.M2, Epidish_TIL10_Macrophages.M2"
#> Keeping only feature: Epidish_TIL10_Macrophages.M2
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_LM22_Macrophages.M2, Epidish_LM22_Macrophages.M2"
#> Keeping only feature: DeconRNASeq_LM22_Macrophages.M2
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_TIL10_Monocytes, CBSX_TIL10_Monocytes"
#> Keeping only feature: Epidish_TIL10_Monocytes
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_BSeqSC.Vanderbilt.scRNAseq_NK.cells, CBSX_BSeqSC.Vanderbilt.scRNAseq_NK.cells, DWLS_BSeqSC.Vanderbilt.scRNAseq_NK.cells, DeconRNASeq_CBSX.Vanderbilt.scRNAseq_NK.cells, DeconRNASeq_MOMF.Vanderbilt.scRNAseq_NK.cells, MuSic_Vanderbilt_NK.cells"
#> Keeping only feature: DWLS_BSeqSC.Vanderbilt.scRNAseq_NK.cells
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_CBSX.Melanoma.scRNAseq_NK.cells, CBSX_CBSX.Melanoma.scRNAseq_NK.cells, DWLS_CBSX.Melanoma.scRNAseq_NK.cells"
#> Keeping only feature: DWLS_CBSX.Melanoma.scRNAseq_NK.cells
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_CBSX.Vanderbilt.scRNAseq_NK.cells, CBSX_CBSX.Vanderbilt.scRNAseq_NK.cells, DWLS_CBSX.Vanderbilt.scRNAseq_NK.cells, DeconRNASeq_DWLS.Vanderbilt.scRNAseq_NK.cells, Epidish_DWLS.Vanderbilt.scRNAseq_NK.cells, CBSX_DWLS.Vanderbilt.scRNAseq_NK.cells, DWLS_DWLS.Vanderbilt.scRNAseq_NK.cells, Epidish_MOMF.Vanderbilt.scRNAseq_NK.cells, CBSX_MOMF.Vanderbilt.scRNAseq_NK.cells"
#> Keeping only feature: DWLS_CBSX.Vanderbilt.scRNAseq_NK.cells
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_CCLE.TIL10_NK.cells, Epidish_CCLE.TIL10_NK.cells, DeconRNASeq_TIL10_NK.cells"
#> Keeping only feature: DeconRNASeq_TIL10_NK.cells
#> 
#> [1] "Highly correlated features (r>0.9): CBSX_CCLE.TIL10_NK.cells, DWLS_CCLE.TIL10_NK.cells, Epidish_TIL10_NK.cells, CBSX_TIL10_NK.cells"
#> Keeping only feature: Epidish_TIL10_NK.cells
#> 
#> [1] "Highly correlated features (r>0.9): DWLS_MOMF.Vanderbilt.scRNAseq_NK.cells, BayesPrism_Vanderbilt_NK.cells, Bisque_Vanderbilt_NK.cells, SCDC_Vanderbilt_NK.cells"
#> Keeping only feature: Bisque_Vanderbilt_NK.cells
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_LM22_NK.activated, CBSX_LM22_NK.activated, DWLS_LM22_NK.activated"
#> Keeping only feature: DWLS_LM22_NK.activated
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_CBSX.NSCLC.PBMCs.scRNAseq_NKT.cells, CBSX_CBSX.NSCLC.PBMCs.scRNAseq_NKT.cells, DWLS_CBSX.NSCLC.PBMCs.scRNAseq_NKT.cells"
#> Keeping only feature: DWLS_CBSX.NSCLC.PBMCs.scRNAseq_NKT.cells
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_BPRNACan_CD4.cells, DeconRNASeq_BPRNACan3DProMet_CD4.cells, DeconRNASeq_BPRNACanProMet_CD4.cells"
#> Keeping only feature: DeconRNASeq_BPRNACanProMet_CD4.cells
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_BPRNACan_CD4.cells, CBSX_BPRNACan_CD4.cells, DWLS_BPRNACan_CD4.cells, Epidish_BPRNACan3DProMet_CD4.cells, Epidish_BPRNACanProMet_CD4.cells"
#> Keeping only feature: DWLS_BPRNACan_CD4.cells
#> 
#> [1] "Highly correlated features (r>0.9): CBSX_BPRNACan3DProMet_CD4.cells, DWLS_BPRNACan3DProMet_CD4.cells, CBSX_BPRNACanProMet_CD4.cells"
#> Keeping only feature: CBSX_BPRNACanProMet_CD4.cells
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_BSeqSC.Vanderbilt.scRNAseq_CD4.cells, CBSX_BSeqSC.Vanderbilt.scRNAseq_CD4.cells, DWLS_BSeqSC.Vanderbilt.scRNAseq_CD4.cells, Epidish_CBSX.Vanderbilt.scRNAseq_CD4.cells, CBSX_CBSX.Vanderbilt.scRNAseq_CD4.cells, DWLS_CBSX.Vanderbilt.scRNAseq_CD4.cells, Epidish_DWLS.Vanderbilt.scRNAseq_CD4.cells, CBSX_DWLS.Vanderbilt.scRNAseq_CD4.cells, DWLS_DWLS.Vanderbilt.scRNAseq_CD4.cells, DWLS_MOMF.Vanderbilt.scRNAseq_CD4.cells"
#> Keeping only feature: DWLS_BSeqSC.Vanderbilt.scRNAseq_CD4.cells
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_CBSX.HNSCC.scRNAseq_CD4.cells, CBSX_CBSX.HNSCC.scRNAseq_CD4.cells, DWLS_CBSX.HNSCC.scRNAseq_CD4.cells"
#> Keeping only feature: DWLS_CBSX.HNSCC.scRNAseq_CD4.cells
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_CBSX.Melanoma.scRNAseq_CD4.cells, CBSX_CBSX.Melanoma.scRNAseq_CD4.cells, DWLS_CBSX.Melanoma.scRNAseq_CD4.cells"
#> Keeping only feature: DWLS_CBSX.Melanoma.scRNAseq_CD4.cells
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_CBSX.NSCLC.PBMCs.scRNAseq_CD4.cells, CBSX_CBSX.NSCLC.PBMCs.scRNAseq_CD4.cells, DWLS_CBSX.NSCLC.PBMCs.scRNAseq_CD4.cells"
#> Keeping only feature: DWLS_CBSX.NSCLC.PBMCs.scRNAseq_CD4.cells
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_CBSX.Vanderbilt.scRNAseq_CD4.cells, DeconRNASeq_DWLS.Vanderbilt.scRNAseq_CD4.cells, DeconRNASeq_MOMF.Vanderbilt.scRNAseq_CD4.cells, Bisque_Vanderbilt_CD4.cells"
#> Keeping only feature: DeconRNASeq_MOMF.Vanderbilt.scRNAseq_CD4.cells
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_CCLE.TIL10_CD4.cells, DeconRNASeq_TIL10_CD4.cells"
#> Keeping only feature: DeconRNASeq_CCLE.TIL10_CD4.cells
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_CCLE.TIL10_CD4.cells, DWLS_CCLE.TIL10_CD4.cells"
#> Keeping only feature: Epidish_CCLE.TIL10_CD4.cells
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_TIL10_CD4.cells, DWLS_TIL10_CD4.cells"
#> Keeping only feature: Epidish_TIL10_CD4.cells
#> 
#> [1] "Highly correlated features (r>0.9): BayesPrism_Vanderbilt_CD4.cells, MuSic_Vanderbilt_CD4.cells"
#> Keeping only feature: BayesPrism_Vanderbilt_CD4.cells
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_LM22_CD4.memory.activated, CBSX_LM22_CD4.memory.activated, DWLS_LM22_CD4.memory.activated"
#> Keeping only feature: DWLS_LM22_CD4.memory.activated
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_LM22_CD4.memory.resting, DWLS_LM22_CD4.memory.resting"
#> Keeping only feature: DeconRNASeq_LM22_CD4.memory.resting
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_LM22_CD4.memory.resting, CBSX_LM22_CD4.memory.resting"
#> Keeping only feature: Epidish_LM22_CD4.memory.resting
#> 
#> [1] "Highly correlated features (r>0.9): Quantiseq_CD8.cells, DeconRNASeq_CCLE.TIL10_CD8.cells, Epidish_CCLE.TIL10_CD8.cells, CBSX_CCLE.TIL10_CD8.cells, DWLS_CCLE.TIL10_CD8.cells, DeconRNASeq_TIL10_CD8.cells, Epidish_TIL10_CD8.cells, CBSX_TIL10_CD8.cells"
#> Keeping only feature: Epidish_TIL10_CD8.cells
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_BPRNACan_CD8.cells, DeconRNASeq_BPRNACan3DProMet_CD8.cells, DeconRNASeq_BPRNACanProMet_CD8.cells"
#> Keeping only feature: DeconRNASeq_BPRNACanProMet_CD8.cells
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_BPRNACan_CD8.cells, CBSX_BPRNACan_CD8.cells, DWLS_BPRNACan_CD8.cells, Epidish_BPRNACan3DProMet_CD8.cells, CBSX_BPRNACan3DProMet_CD8.cells, DWLS_BPRNACan3DProMet_CD8.cells, Epidish_BPRNACanProMet_CD8.cells, CBSX_BPRNACanProMet_CD8.cells, DWLS_BPRNACanProMet_CD8.cells"
#> Keeping only feature: DWLS_BPRNACan_CD8.cells
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_BSeqSC.Vanderbilt.scRNAseq_CD8.cells, DWLS_BSeqSC.Vanderbilt.scRNAseq_CD8.cells"
#> Keeping only feature: Epidish_BSeqSC.Vanderbilt.scRNAseq_CD8.cells
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_CBSX.HNSCC.scRNAseq_CD8.cells, Epidish_CBSX.HNSCC.scRNAseq_CD8.cells, CBSX_CBSX.HNSCC.scRNAseq_CD8.cells, Epidish_CBSX.NSCLC.PBMCs.scRNAseq_CD8.cells, CBSX_CBSX.NSCLC.PBMCs.scRNAseq_CD8.cells"
#> Keeping only feature: CBSX_CBSX.HNSCC.scRNAseq_CD8.cells
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_CBSX.Melanoma.scRNAseq_CD8.cells, CBSX_CBSX.Melanoma.scRNAseq_CD8.cells, DWLS_CBSX.Melanoma.scRNAseq_CD8.cells"
#> Keeping only feature: DWLS_CBSX.Melanoma.scRNAseq_CD8.cells
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_CBSX.Vanderbilt.scRNAseq_CD8.cells, CBSX_CBSX.Vanderbilt.scRNAseq_CD8.cells, DWLS_CBSX.Vanderbilt.scRNAseq_CD8.cells"
#> Keeping only feature: DWLS_CBSX.Vanderbilt.scRNAseq_CD8.cells
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_DWLS.Vanderbilt.scRNAseq_CD8.cells, Epidish_DWLS.Vanderbilt.scRNAseq_CD8.cells, CBSX_DWLS.Vanderbilt.scRNAseq_CD8.cells, DWLS_DWLS.Vanderbilt.scRNAseq_CD8.cells"
#> Keeping only feature: CBSX_DWLS.Vanderbilt.scRNAseq_CD8.cells
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_LM22_CD8.cells, CBSX_LM22_CD8.cells, DWLS_LM22_CD8.cells"
#> Keeping only feature: DWLS_LM22_CD8.cells
#> 
#> [1] "Highly correlated features (r>0.9): DWLS_MOMF.Vanderbilt.scRNAseq_CD8.cells, BayesPrism_Vanderbilt_CD8.cells, MuSic_Vanderbilt_CD8.cells"
#> Keeping only feature: MuSic_Vanderbilt_CD8.cells
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_CCLE.TIL10_T.cells.regulatory, DeconRNASeq_TIL10_T.cells.regulatory"
#> Keeping only feature: DeconRNASeq_CCLE.TIL10_T.cells.regulatory
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_CCLE.TIL10_T.cells.regulatory, DWLS_CCLE.TIL10_T.cells.regulatory"
#> Keeping only feature: Epidish_CCLE.TIL10_T.cells.regulatory
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_TIL10_T.cells.regulatory, DWLS_TIL10_T.cells.regulatory"
#> Keeping only feature: Epidish_TIL10_T.cells.regulatory
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_BSeqSC.Vanderbilt.scRNAseq_T.cells.regulatory, CBSX_BSeqSC.Vanderbilt.scRNAseq_T.cells.regulatory, DWLS_BSeqSC.Vanderbilt.scRNAseq_T.cells.regulatory"
#> Keeping only feature: DWLS_BSeqSC.Vanderbilt.scRNAseq_T.cells.regulatory
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_CBSX.Vanderbilt.scRNAseq_T.cells.regulatory, CBSX_CBSX.Vanderbilt.scRNAseq_T.cells.regulatory, DWLS_CBSX.Vanderbilt.scRNAseq_T.cells.regulatory"
#> Keeping only feature: DWLS_CBSX.Vanderbilt.scRNAseq_T.cells.regulatory
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_DWLS.Vanderbilt.scRNAseq_T.cells.regulatory, CBSX_MOMF.Vanderbilt.scRNAseq_T.cells.regulatory"
#> Keeping only feature: DeconRNASeq_DWLS.Vanderbilt.scRNAseq_T.cells.regulatory
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_DWLS.Vanderbilt.scRNAseq_T.cells.regulatory, CBSX_DWLS.Vanderbilt.scRNAseq_T.cells.regulatory, DWLS_DWLS.Vanderbilt.scRNAseq_T.cells.regulatory, DWLS_MOMF.Vanderbilt.scRNAseq_T.cells.regulatory"
#> Keeping only feature: DWLS_DWLS.Vanderbilt.scRNAseq_T.cells.regulatory
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_BSeqSC.Vanderbilt.scRNAseq_T.cells.helper, CBSX_BSeqSC.Vanderbilt.scRNAseq_T.cells.helper, DWLS_BSeqSC.Vanderbilt.scRNAseq_T.cells.helper, DeconRNASeq_CBSX.Vanderbilt.scRNAseq_T.cells.helper, Epidish_CBSX.Vanderbilt.scRNAseq_T.cells.helper, CBSX_CBSX.Vanderbilt.scRNAseq_T.cells.helper, DWLS_CBSX.Vanderbilt.scRNAseq_T.cells.helper, DeconRNASeq_DWLS.Vanderbilt.scRNAseq_T.cells.helper, Epidish_DWLS.Vanderbilt.scRNAseq_T.cells.helper, CBSX_DWLS.Vanderbilt.scRNAseq_T.cells.helper, DWLS_DWLS.Vanderbilt.scRNAseq_T.cells.helper, DeconRNASeq_LM22_T.cells.helper, DWLS_MOMF.Vanderbilt.scRNAseq_T.cells.helper, BayesPrism_Vanderbilt_T.cells.helper, Bisque_Vanderbilt_T.cells.helper"
#> Keeping only feature: Bisque_Vanderbilt_T.cells.helper
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_LM22_T.cells.helper, CBSX_LM22_T.cells.helper, DWLS_LM22_T.cells.helper"
#> Keeping only feature: DWLS_LM22_T.cells.helper
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_MOMF.Vanderbilt.scRNAseq_T.cells.helper, MuSic_Vanderbilt_T.cells.helper, SCDC_Vanderbilt_T.cells.helper"
#> Keeping only feature: SCDC_Vanderbilt_T.cells.helper
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_MOMF.Vanderbilt.scRNAseq_T.cells.helper, CBSX_MOMF.Vanderbilt.scRNAseq_T.cells.helper"
#> Keeping only feature: Epidish_MOMF.Vanderbilt.scRNAseq_T.cells.helper
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_CCLE.TIL10_Dendritic.cells, DeconRNASeq_TIL10_Dendritic.cells"
#> Keeping only feature: DeconRNASeq_CCLE.TIL10_Dendritic.cells
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_BPRNACan_Cancer, DeconRNASeq_BPRNACan3DProMet_Cancer, CBSX_BPRNACan3DProMet_Cancer, DeconRNASeq_BPRNACanProMet_Cancer"
#> Keeping only feature: CBSX_BPRNACan3DProMet_Cancer
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_BPRNACan_Cancer, CBSX_BPRNACan_Cancer, DWLS_BPRNACan_Cancer, Epidish_BPRNACan3DProMet_Cancer, DWLS_BPRNACan3DProMet_Cancer, Epidish_BPRNACanProMet_Cancer, CBSX_BPRNACanProMet_Cancer, DWLS_BPRNACanProMet_Cancer"
#> Keeping only feature: CBSX_BPRNACanProMet_Cancer
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_CCLE.TIL10_Cancer, CBSX_CCLE.TIL10_Cancer, DWLS_CCLE.TIL10_Cancer"
#> Keeping only feature: DWLS_CCLE.TIL10_Cancer
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_BSeqSC.Vanderbilt.scRNAseq_Endothelial, DWLS_BSeqSC.Vanderbilt.scRNAseq_Endothelial, DeconRNASeq_CBSX.HNSCC.scRNAseq_Endothelial, DWLS_CBSX.Melanoma.scRNAseq_Endothelial, DeconRNASeq_CBSX.Vanderbilt.scRNAseq_Endothelial, DWLS_CBSX.Vanderbilt.scRNAseq_Endothelial, DeconRNASeq_DWLS.Vanderbilt.scRNAseq_Endothelial, Epidish_DWLS.Vanderbilt.scRNAseq_Endothelial, CBSX_DWLS.Vanderbilt.scRNAseq_Endothelial, DWLS_DWLS.Vanderbilt.scRNAseq_Endothelial, CBSX_MOMF.Vanderbilt.scRNAseq_Endothelial, BayesPrism_Vanderbilt_Endothelial, SCDC_Vanderbilt_Endothelial"
#> Keeping only feature: DeconRNASeq_CBSX.HNSCC.scRNAseq_Endothelial
#> 
#> [1] "Highly correlated features (r>0.9): CBSX_BSeqSC.Vanderbilt.scRNAseq_Endothelial, Epidish_MOMF.Vanderbilt.scRNAseq_Endothelial"
#> Keeping only feature: CBSX_BSeqSC.Vanderbilt.scRNAseq_Endothelial
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_BSeqSC.Vanderbilt.scRNAseq_Fibroblasts, Epidish_BSeqSC.Vanderbilt.scRNAseq_Fibroblasts, CBSX_BSeqSC.Vanderbilt.scRNAseq_Fibroblasts, DWLS_BSeqSC.Vanderbilt.scRNAseq_Fibroblasts"
#> Keeping only feature: CBSX_BSeqSC.Vanderbilt.scRNAseq_Fibroblasts
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_CBSX.Vanderbilt.scRNAseq_Fibroblasts, Epidish_CBSX.Vanderbilt.scRNAseq_Fibroblasts, CBSX_CBSX.Vanderbilt.scRNAseq_Fibroblasts, DWLS_CBSX.Vanderbilt.scRNAseq_Fibroblasts, DeconRNASeq_DWLS.Vanderbilt.scRNAseq_Fibroblasts, Epidish_DWLS.Vanderbilt.scRNAseq_Fibroblasts, CBSX_DWLS.Vanderbilt.scRNAseq_Fibroblasts, DWLS_DWLS.Vanderbilt.scRNAseq_Fibroblasts, DWLS_MOMF.Vanderbilt.scRNAseq_Fibroblasts, Bisque_Vanderbilt_Fibroblasts"
#> Keeping only feature: CBSX_CBSX.Vanderbilt.scRNAseq_Fibroblasts
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_MOMF.Vanderbilt.scRNAseq_Fibroblasts, CBSX_MOMF.Vanderbilt.scRNAseq_Fibroblasts, AutogeneS_Vanderbilt_Fibroblasts, BayesPrism_Vanderbilt_Fibroblasts, MuSic_Vanderbilt_Fibroblasts, SCDC_Vanderbilt_Fibroblasts"
#> Keeping only feature: AutogeneS_Vanderbilt_Fibroblasts
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_BSeqSC.Vanderbilt.scRNAseq_Mast.cells, CBSX_BSeqSC.Vanderbilt.scRNAseq_Mast.cells, DWLS_BSeqSC.Vanderbilt.scRNAseq_Mast.cells, Epidish_DWLS.Vanderbilt.scRNAseq_Mast.cells, DeconRNASeq_MOMF.Vanderbilt.scRNAseq_Mast.cells"
#> Keeping only feature: DWLS_BSeqSC.Vanderbilt.scRNAseq_Mast.cells
#> 
#> [1] "Highly correlated features (r>0.9): DWLS_CBSX.Vanderbilt.scRNAseq_Mast.cells, DeconRNASeq_DWLS.Vanderbilt.scRNAseq_Mast.cells, CBSX_DWLS.Vanderbilt.scRNAseq_Mast.cells, DWLS_DWLS.Vanderbilt.scRNAseq_Mast.cells, Epidish_MOMF.Vanderbilt.scRNAseq_Mast.cells, CBSX_MOMF.Vanderbilt.scRNAseq_Mast.cells, DWLS_MOMF.Vanderbilt.scRNAseq_Mast.cells, BayesPrism_Vanderbilt_Mast.cells, Bisque_Vanderbilt_Mast.cells, SCDC_Vanderbilt_Mast.cells"
#> Keeping only feature: CBSX_DWLS.Vanderbilt.scRNAseq_Mast.cells
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_LM22_Mast.activated.cells, CBSX_LM22_Mast.activated.cells, DWLS_LM22_Mast.activated.cells"
#> Keeping only feature: DWLS_LM22_Mast.activated.cells
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_CBSX.HNSCC.scRNAseq_CAF, Epidish_CBSX.HNSCC.scRNAseq_CAF, CBSX_CBSX.HNSCC.scRNAseq_CAF, DWLS_CBSX.HNSCC.scRNAseq_CAF, DeconRNASeq_CBSX.Melanoma.scRNAseq_CAF, Epidish_CBSX.Melanoma.scRNAseq_CAF, CBSX_CBSX.Melanoma.scRNAseq_CAF, DWLS_CBSX.Melanoma.scRNAseq_CAF"
#> Keeping only feature: CBSX_CBSX.Melanoma.scRNAseq_CAF
#> 
#> Warning: Deconvolution features with less than two columns for subgrouping (skipping)
#> Warning: Deconvolution features with less than two columns for subgrouping (skipping)
#> Warning: Deconvolution features with less than two columns for subgrouping (skipping)
#> Warning: Deconvolution features with less than two columns for subgrouping (skipping)
#> Warning: Deconvolution features with less than two columns for subgrouping (skipping)
#> Warning: Deconvolution features with less than two columns for subgrouping (skipping)
#> Warning: Deconvolution features with less than two columns for subgrouping (skipping)
#> Warning: Deconvolution features with less than two columns for subgrouping (skipping)
#> Warning: Deconvolution features with less than two columns for subgrouping (skipping)
#> Warning: Deconvolution features with less than two columns for subgrouping (skipping)
#> Warning: Deconvolution features with less than two columns for subgrouping (skipping)
#> Warning: Deconvolution features with less than two columns for subgrouping (skipping)
#> Warning: Deconvolution features with less than two columns for subgrouping (skipping)
#> Warning: Deconvolution features with less than two columns for subgrouping (skipping)
#> Warning: Deconvolution features with less than two columns for subgrouping (skipping)
#> Deconvolution features subgroupped

processed_deconvolution = compute.deconvolution.analysis(deconvolution, cells_extra = "mesenchymal")
#> Removing features with high zero number 90%...............................................................
#> 
#> Checking distribution of variances...............................................................
#> 
#> Chosen threshold is: 0.0002698755 
#> 
#> Saving variance distribution plot in Results/ folder
#> 
#> Removing 157 features with variance across samples below this threshold...............................................................
#> 
#> Splitting deconvolution features per cell type...............................................................
#> 
#> Finding group of features with high correlation between each other...............................................................
#> 
#> [1] "Highly correlated features (r>0.9): Quantiseq_B.cells, DeconRNASeq_CBSX.HNSCC.scRNAseq_B.cells, DWLS_CBSX.HNSCC.scRNAseq_B.cells, DeconRNASeq_CBSX.Melanoma.scRNAseq_B.cells, Epidish_CBSX.Melanoma.scRNAseq_B.cells, CBSX_CBSX.Melanoma.scRNAseq_B.cells, DWLS_CBSX.Melanoma.scRNAseq_B.cells, DeconRNASeq_CCLE.TIL10_B.cells, Epidish_CCLE.TIL10_B.cells, CBSX_CCLE.TIL10_B.cells, DWLS_CCLE.TIL10_B.cells, DeconRNASeq_TIL10_B.cells, Epidish_TIL10_B.cells, CBSX_TIL10_B.cells, DWLS_TIL10_B.cells"
#> Keeping only feature: DeconRNASeq_CCLE.TIL10_B.cells
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_BPRNACan_B.cells, Epidish_BPRNACan_B.cells, CBSX_BPRNACan_B.cells, DWLS_BPRNACan_B.cells, DeconRNASeq_BPRNACan3DProMet_B.cells, Epidish_BPRNACan3DProMet_B.cells, CBSX_BPRNACan3DProMet_B.cells, DWLS_BPRNACan3DProMet_B.cells, DeconRNASeq_BPRNACanProMet_B.cells, Epidish_BPRNACanProMet_B.cells, CBSX_BPRNACanProMet_B.cells, DWLS_BPRNACanProMet_B.cells, DeconRNASeq_BSeqSC.Vanderbilt.scRNAseq_B.cells, Epidish_BSeqSC.Vanderbilt.scRNAseq_B.cells, DWLS_BSeqSC.Vanderbilt.scRNAseq_B.cells, Epidish_DWLS.Vanderbilt.scRNAseq_B.cells, CBSX_DWLS.Vanderbilt.scRNAseq_B.cells, DWLS_DWLS.Vanderbilt.scRNAseq_B.cells, DeconRNASeq_MOMF.Vanderbilt.scRNAseq_B.cells, DWLS_MOMF.Vanderbilt.scRNAseq_B.cells, BayesPrism_Vanderbilt_B.cells, Bisque_Vanderbilt_B.cells, MuSic_Vanderbilt_B.cells, SCDC_Vanderbilt_B.cells"
#> Keeping only feature: DWLS_BPRNACanProMet_B.cells
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_CBSX.HNSCC.scRNAseq_B.cells, CBSX_CBSX.HNSCC.scRNAseq_B.cells"
#> Keeping only feature: Epidish_CBSX.HNSCC.scRNAseq_B.cells
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_CBSX.NSCLC.PBMCs.scRNAseq_B.cells, Epidish_CBSX.NSCLC.PBMCs.scRNAseq_B.cells, CBSX_CBSX.NSCLC.PBMCs.scRNAseq_B.cells, DWLS_CBSX.NSCLC.PBMCs.scRNAseq_B.cells"
#> Keeping only feature: CBSX_CBSX.NSCLC.PBMCs.scRNAseq_B.cells
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_CBSX.Vanderbilt.scRNAseq_B.cells, Epidish_CBSX.Vanderbilt.scRNAseq_B.cells, CBSX_CBSX.Vanderbilt.scRNAseq_B.cells"
#> Keeping only feature: CBSX_CBSX.Vanderbilt.scRNAseq_B.cells
#> 
#> [1] "Highly correlated features (r>0.9): DWLS_CBSX.Vanderbilt.scRNAseq_B.cells, DeconRNASeq_DWLS.Vanderbilt.scRNAseq_B.cells"
#> Keeping only feature: DeconRNASeq_DWLS.Vanderbilt.scRNAseq_B.cells
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_LM22_B.memory.cells, Epidish_LM22_B.memory.cells, CBSX_LM22_B.memory.cells, DWLS_LM22_B.memory.cells"
#> Keeping only feature: DWLS_LM22_B.memory.cells
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_CBSX.HNSCC.scRNAseq_Macrophages.cells, Epidish_CBSX.HNSCC.scRNAseq_Macrophages.cells, CBSX_CBSX.HNSCC.scRNAseq_Macrophages.cells, DWLS_CBSX.HNSCC.scRNAseq_Macrophages.cells, Epidish_CBSX.Melanoma.scRNAseq_Macrophages.cells, CBSX_CBSX.Melanoma.scRNAseq_Macrophages.cells, DWLS_CBSX.Melanoma.scRNAseq_Macrophages.cells"
#> Keeping only feature: DeconRNASeq_CBSX.HNSCC.scRNAseq_Macrophages.cells
#> 
#> [1] "Highly correlated features (r>0.9): Quantiseq_Macrophages.M1, DeconRNASeq_CCLE.TIL10_Macrophages.M1, Epidish_CCLE.TIL10_Macrophages.M1, DWLS_CCLE.TIL10_Macrophages.M1, DeconRNASeq_TIL10_Macrophages.M1, Epidish_TIL10_Macrophages.M1, DWLS_TIL10_Macrophages.M1"
#> Keeping only feature: Epidish_CCLE.TIL10_Macrophages.M1
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_BPRNACan_Macrophages.M2, DeconRNASeq_BPRNACan3DProMet_Macrophages.M2, DeconRNASeq_BPRNACanProMet_Macrophages.M2"
#> Keeping only feature: DeconRNASeq_BPRNACan_Macrophages.M2
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_CCLE.TIL10_Macrophages.M2, DWLS_CCLE.TIL10_Macrophages.M2, Epidish_TIL10_Macrophages.M2"
#> Keeping only feature: DWLS_CCLE.TIL10_Macrophages.M2
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_LM22_Macrophages.M2, Epidish_LM22_Macrophages.M2"
#> Keeping only feature: DeconRNASeq_LM22_Macrophages.M2
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_TIL10_Monocytes, CBSX_TIL10_Monocytes"
#> Keeping only feature: CBSX_TIL10_Monocytes
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_BSeqSC.Vanderbilt.scRNAseq_NK.cells, CBSX_BSeqSC.Vanderbilt.scRNAseq_NK.cells, DWLS_BSeqSC.Vanderbilt.scRNAseq_NK.cells, DeconRNASeq_CBSX.Vanderbilt.scRNAseq_NK.cells, DeconRNASeq_MOMF.Vanderbilt.scRNAseq_NK.cells, MuSic_Vanderbilt_NK.cells"
#> Keeping only feature: DWLS_BSeqSC.Vanderbilt.scRNAseq_NK.cells
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_CBSX.Melanoma.scRNAseq_NK.cells, CBSX_CBSX.Melanoma.scRNAseq_NK.cells, DWLS_CBSX.Melanoma.scRNAseq_NK.cells"
#> Keeping only feature: DWLS_CBSX.Melanoma.scRNAseq_NK.cells
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_CBSX.Vanderbilt.scRNAseq_NK.cells, CBSX_CBSX.Vanderbilt.scRNAseq_NK.cells, DWLS_CBSX.Vanderbilt.scRNAseq_NK.cells, DeconRNASeq_DWLS.Vanderbilt.scRNAseq_NK.cells, Epidish_DWLS.Vanderbilt.scRNAseq_NK.cells, CBSX_DWLS.Vanderbilt.scRNAseq_NK.cells, DWLS_DWLS.Vanderbilt.scRNAseq_NK.cells, Epidish_MOMF.Vanderbilt.scRNAseq_NK.cells, CBSX_MOMF.Vanderbilt.scRNAseq_NK.cells"
#> Keeping only feature: CBSX_CBSX.Vanderbilt.scRNAseq_NK.cells
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_CCLE.TIL10_NK.cells, Epidish_CCLE.TIL10_NK.cells, DeconRNASeq_TIL10_NK.cells"
#> Keeping only feature: Epidish_CCLE.TIL10_NK.cells
#> 
#> [1] "Highly correlated features (r>0.9): CBSX_CCLE.TIL10_NK.cells, DWLS_CCLE.TIL10_NK.cells, Epidish_TIL10_NK.cells, CBSX_TIL10_NK.cells"
#> Keeping only feature: DWLS_CCLE.TIL10_NK.cells
#> 
#> [1] "Highly correlated features (r>0.9): DWLS_MOMF.Vanderbilt.scRNAseq_NK.cells, BayesPrism_Vanderbilt_NK.cells, Bisque_Vanderbilt_NK.cells, SCDC_Vanderbilt_NK.cells"
#> Keeping only feature: SCDC_Vanderbilt_NK.cells
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_LM22_NK.activated, CBSX_LM22_NK.activated, DWLS_LM22_NK.activated"
#> Keeping only feature: Epidish_LM22_NK.activated
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_CBSX.NSCLC.PBMCs.scRNAseq_NKT.cells, CBSX_CBSX.NSCLC.PBMCs.scRNAseq_NKT.cells, DWLS_CBSX.NSCLC.PBMCs.scRNAseq_NKT.cells"
#> Keeping only feature: CBSX_CBSX.NSCLC.PBMCs.scRNAseq_NKT.cells
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_BPRNACan_CD4.cells, DeconRNASeq_BPRNACan3DProMet_CD4.cells, DeconRNASeq_BPRNACanProMet_CD4.cells"
#> Keeping only feature: DeconRNASeq_BPRNACan_CD4.cells
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_BPRNACan_CD4.cells, CBSX_BPRNACan_CD4.cells, DWLS_BPRNACan_CD4.cells, Epidish_BPRNACan3DProMet_CD4.cells, Epidish_BPRNACanProMet_CD4.cells"
#> Keeping only feature: DWLS_BPRNACan_CD4.cells
#> 
#> [1] "Highly correlated features (r>0.9): CBSX_BPRNACan3DProMet_CD4.cells, DWLS_BPRNACan3DProMet_CD4.cells, CBSX_BPRNACanProMet_CD4.cells"
#> Keeping only feature: CBSX_BPRNACanProMet_CD4.cells
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_BSeqSC.Vanderbilt.scRNAseq_CD4.cells, CBSX_BSeqSC.Vanderbilt.scRNAseq_CD4.cells, DWLS_BSeqSC.Vanderbilt.scRNAseq_CD4.cells, Epidish_CBSX.Vanderbilt.scRNAseq_CD4.cells, CBSX_CBSX.Vanderbilt.scRNAseq_CD4.cells, DWLS_CBSX.Vanderbilt.scRNAseq_CD4.cells, Epidish_DWLS.Vanderbilt.scRNAseq_CD4.cells, CBSX_DWLS.Vanderbilt.scRNAseq_CD4.cells, DWLS_DWLS.Vanderbilt.scRNAseq_CD4.cells, DWLS_MOMF.Vanderbilt.scRNAseq_CD4.cells"
#> Keeping only feature: Epidish_DWLS.Vanderbilt.scRNAseq_CD4.cells
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_CBSX.HNSCC.scRNAseq_CD4.cells, CBSX_CBSX.HNSCC.scRNAseq_CD4.cells, DWLS_CBSX.HNSCC.scRNAseq_CD4.cells"
#> Keeping only feature: DWLS_CBSX.HNSCC.scRNAseq_CD4.cells
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_CBSX.Melanoma.scRNAseq_CD4.cells, CBSX_CBSX.Melanoma.scRNAseq_CD4.cells, DWLS_CBSX.Melanoma.scRNAseq_CD4.cells"
#> Keeping only feature: Epidish_CBSX.Melanoma.scRNAseq_CD4.cells
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_CBSX.NSCLC.PBMCs.scRNAseq_CD4.cells, CBSX_CBSX.NSCLC.PBMCs.scRNAseq_CD4.cells, DWLS_CBSX.NSCLC.PBMCs.scRNAseq_CD4.cells"
#> Keeping only feature: Epidish_CBSX.NSCLC.PBMCs.scRNAseq_CD4.cells
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_CBSX.Vanderbilt.scRNAseq_CD4.cells, DeconRNASeq_DWLS.Vanderbilt.scRNAseq_CD4.cells, DeconRNASeq_MOMF.Vanderbilt.scRNAseq_CD4.cells, Bisque_Vanderbilt_CD4.cells"
#> Keeping only feature: DeconRNASeq_DWLS.Vanderbilt.scRNAseq_CD4.cells
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_CCLE.TIL10_CD4.cells, DeconRNASeq_TIL10_CD4.cells"
#> Keeping only feature: DeconRNASeq_TIL10_CD4.cells
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_CCLE.TIL10_CD4.cells, DWLS_CCLE.TIL10_CD4.cells"
#> Keeping only feature: DWLS_CCLE.TIL10_CD4.cells
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_TIL10_CD4.cells, DWLS_TIL10_CD4.cells"
#> Keeping only feature: Epidish_TIL10_CD4.cells
#> 
#> [1] "Highly correlated features (r>0.9): BayesPrism_Vanderbilt_CD4.cells, MuSic_Vanderbilt_CD4.cells"
#> Keeping only feature: BayesPrism_Vanderbilt_CD4.cells
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_LM22_CD4.memory.activated, CBSX_LM22_CD4.memory.activated, DWLS_LM22_CD4.memory.activated"
#> Keeping only feature: Epidish_LM22_CD4.memory.activated
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_LM22_CD4.memory.resting, DWLS_LM22_CD4.memory.resting"
#> Keeping only feature: DeconRNASeq_LM22_CD4.memory.resting
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_LM22_CD4.memory.resting, CBSX_LM22_CD4.memory.resting"
#> Keeping only feature: CBSX_LM22_CD4.memory.resting
#> 
#> [1] "Highly correlated features (r>0.9): Quantiseq_CD8.cells, DeconRNASeq_CCLE.TIL10_CD8.cells, Epidish_CCLE.TIL10_CD8.cells, CBSX_CCLE.TIL10_CD8.cells, DWLS_CCLE.TIL10_CD8.cells, DeconRNASeq_TIL10_CD8.cells, Epidish_TIL10_CD8.cells, CBSX_TIL10_CD8.cells"
#> Keeping only feature: Quantiseq_CD8.cells
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_BPRNACan_CD8.cells, DeconRNASeq_BPRNACan3DProMet_CD8.cells, DeconRNASeq_BPRNACanProMet_CD8.cells"
#> Keeping only feature: DeconRNASeq_BPRNACanProMet_CD8.cells
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_BPRNACan_CD8.cells, CBSX_BPRNACan_CD8.cells, DWLS_BPRNACan_CD8.cells, Epidish_BPRNACan3DProMet_CD8.cells, CBSX_BPRNACan3DProMet_CD8.cells, DWLS_BPRNACan3DProMet_CD8.cells, Epidish_BPRNACanProMet_CD8.cells, CBSX_BPRNACanProMet_CD8.cells, DWLS_BPRNACanProMet_CD8.cells"
#> Keeping only feature: CBSX_BPRNACanProMet_CD8.cells
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_BSeqSC.Vanderbilt.scRNAseq_CD8.cells, DWLS_BSeqSC.Vanderbilt.scRNAseq_CD8.cells"
#> Keeping only feature: DWLS_BSeqSC.Vanderbilt.scRNAseq_CD8.cells
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_CBSX.HNSCC.scRNAseq_CD8.cells, Epidish_CBSX.HNSCC.scRNAseq_CD8.cells, CBSX_CBSX.HNSCC.scRNAseq_CD8.cells, Epidish_CBSX.NSCLC.PBMCs.scRNAseq_CD8.cells, CBSX_CBSX.NSCLC.PBMCs.scRNAseq_CD8.cells"
#> Keeping only feature: Epidish_CBSX.NSCLC.PBMCs.scRNAseq_CD8.cells
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_CBSX.Melanoma.scRNAseq_CD8.cells, CBSX_CBSX.Melanoma.scRNAseq_CD8.cells, DWLS_CBSX.Melanoma.scRNAseq_CD8.cells"
#> Keeping only feature: DWLS_CBSX.Melanoma.scRNAseq_CD8.cells
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_CBSX.Vanderbilt.scRNAseq_CD8.cells, CBSX_CBSX.Vanderbilt.scRNAseq_CD8.cells, DWLS_CBSX.Vanderbilt.scRNAseq_CD8.cells"
#> Keeping only feature: DWLS_CBSX.Vanderbilt.scRNAseq_CD8.cells
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_DWLS.Vanderbilt.scRNAseq_CD8.cells, Epidish_DWLS.Vanderbilt.scRNAseq_CD8.cells, CBSX_DWLS.Vanderbilt.scRNAseq_CD8.cells, DWLS_DWLS.Vanderbilt.scRNAseq_CD8.cells"
#> Keeping only feature: DeconRNASeq_DWLS.Vanderbilt.scRNAseq_CD8.cells
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_LM22_CD8.cells, CBSX_LM22_CD8.cells, DWLS_LM22_CD8.cells"
#> Keeping only feature: CBSX_LM22_CD8.cells
#> 
#> [1] "Highly correlated features (r>0.9): DWLS_MOMF.Vanderbilt.scRNAseq_CD8.cells, BayesPrism_Vanderbilt_CD8.cells, MuSic_Vanderbilt_CD8.cells"
#> Keeping only feature: BayesPrism_Vanderbilt_CD8.cells
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_CCLE.TIL10_T.cells.regulatory, DeconRNASeq_TIL10_T.cells.regulatory"
#> Keeping only feature: DeconRNASeq_TIL10_T.cells.regulatory
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_CCLE.TIL10_T.cells.regulatory, DWLS_CCLE.TIL10_T.cells.regulatory"
#> Keeping only feature: Epidish_CCLE.TIL10_T.cells.regulatory
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_TIL10_T.cells.regulatory, DWLS_TIL10_T.cells.regulatory"
#> Keeping only feature: DWLS_TIL10_T.cells.regulatory
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_BSeqSC.Vanderbilt.scRNAseq_T.cells.regulatory, CBSX_BSeqSC.Vanderbilt.scRNAseq_T.cells.regulatory, DWLS_BSeqSC.Vanderbilt.scRNAseq_T.cells.regulatory"
#> Keeping only feature: DWLS_BSeqSC.Vanderbilt.scRNAseq_T.cells.regulatory
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_CBSX.Vanderbilt.scRNAseq_T.cells.regulatory, CBSX_CBSX.Vanderbilt.scRNAseq_T.cells.regulatory, DWLS_CBSX.Vanderbilt.scRNAseq_T.cells.regulatory"
#> Keeping only feature: DWLS_CBSX.Vanderbilt.scRNAseq_T.cells.regulatory
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_DWLS.Vanderbilt.scRNAseq_T.cells.regulatory, CBSX_MOMF.Vanderbilt.scRNAseq_T.cells.regulatory"
#> Keeping only feature: CBSX_MOMF.Vanderbilt.scRNAseq_T.cells.regulatory
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_DWLS.Vanderbilt.scRNAseq_T.cells.regulatory, CBSX_DWLS.Vanderbilt.scRNAseq_T.cells.regulatory, DWLS_DWLS.Vanderbilt.scRNAseq_T.cells.regulatory, DWLS_MOMF.Vanderbilt.scRNAseq_T.cells.regulatory"
#> Keeping only feature: DWLS_DWLS.Vanderbilt.scRNAseq_T.cells.regulatory
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_BSeqSC.Vanderbilt.scRNAseq_T.cells.helper, CBSX_BSeqSC.Vanderbilt.scRNAseq_T.cells.helper, DWLS_BSeqSC.Vanderbilt.scRNAseq_T.cells.helper, DeconRNASeq_CBSX.Vanderbilt.scRNAseq_T.cells.helper, Epidish_CBSX.Vanderbilt.scRNAseq_T.cells.helper, CBSX_CBSX.Vanderbilt.scRNAseq_T.cells.helper, DWLS_CBSX.Vanderbilt.scRNAseq_T.cells.helper, DeconRNASeq_DWLS.Vanderbilt.scRNAseq_T.cells.helper, Epidish_DWLS.Vanderbilt.scRNAseq_T.cells.helper, CBSX_DWLS.Vanderbilt.scRNAseq_T.cells.helper, DWLS_DWLS.Vanderbilt.scRNAseq_T.cells.helper, DeconRNASeq_LM22_T.cells.helper, DWLS_MOMF.Vanderbilt.scRNAseq_T.cells.helper, BayesPrism_Vanderbilt_T.cells.helper, Bisque_Vanderbilt_T.cells.helper"
#> Keeping only feature: CBSX_BSeqSC.Vanderbilt.scRNAseq_T.cells.helper
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_LM22_T.cells.helper, CBSX_LM22_T.cells.helper, DWLS_LM22_T.cells.helper"
#> Keeping only feature: Epidish_LM22_T.cells.helper
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_MOMF.Vanderbilt.scRNAseq_T.cells.helper, MuSic_Vanderbilt_T.cells.helper, SCDC_Vanderbilt_T.cells.helper"
#> Keeping only feature: SCDC_Vanderbilt_T.cells.helper
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_MOMF.Vanderbilt.scRNAseq_T.cells.helper, CBSX_MOMF.Vanderbilt.scRNAseq_T.cells.helper"
#> Keeping only feature: Epidish_MOMF.Vanderbilt.scRNAseq_T.cells.helper
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_CCLE.TIL10_Dendritic.cells, DeconRNASeq_TIL10_Dendritic.cells"
#> Keeping only feature: DeconRNASeq_TIL10_Dendritic.cells
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_BPRNACan_Cancer, DeconRNASeq_BPRNACan3DProMet_Cancer, CBSX_BPRNACan3DProMet_Cancer, DeconRNASeq_BPRNACanProMet_Cancer"
#> Keeping only feature: CBSX_BPRNACan3DProMet_Cancer
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_BPRNACan_Cancer, CBSX_BPRNACan_Cancer, DWLS_BPRNACan_Cancer, Epidish_BPRNACan3DProMet_Cancer, DWLS_BPRNACan3DProMet_Cancer, Epidish_BPRNACanProMet_Cancer, CBSX_BPRNACanProMet_Cancer, DWLS_BPRNACanProMet_Cancer"
#> Keeping only feature: Epidish_BPRNACanProMet_Cancer
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_CCLE.TIL10_Cancer, CBSX_CCLE.TIL10_Cancer, DWLS_CCLE.TIL10_Cancer"
#> Keeping only feature: DWLS_CCLE.TIL10_Cancer
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_BSeqSC.Vanderbilt.scRNAseq_Endothelial, DWLS_BSeqSC.Vanderbilt.scRNAseq_Endothelial, DeconRNASeq_CBSX.HNSCC.scRNAseq_Endothelial, DWLS_CBSX.Melanoma.scRNAseq_Endothelial, DeconRNASeq_CBSX.Vanderbilt.scRNAseq_Endothelial, DWLS_CBSX.Vanderbilt.scRNAseq_Endothelial, DeconRNASeq_DWLS.Vanderbilt.scRNAseq_Endothelial, Epidish_DWLS.Vanderbilt.scRNAseq_Endothelial, CBSX_DWLS.Vanderbilt.scRNAseq_Endothelial, DWLS_DWLS.Vanderbilt.scRNAseq_Endothelial, CBSX_MOMF.Vanderbilt.scRNAseq_Endothelial, BayesPrism_Vanderbilt_Endothelial, SCDC_Vanderbilt_Endothelial"
#> Keeping only feature: DeconRNASeq_CBSX.HNSCC.scRNAseq_Endothelial
#> 
#> [1] "Highly correlated features (r>0.9): CBSX_BSeqSC.Vanderbilt.scRNAseq_Endothelial, Epidish_MOMF.Vanderbilt.scRNAseq_Endothelial"
#> Keeping only feature: CBSX_BSeqSC.Vanderbilt.scRNAseq_Endothelial
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_BSeqSC.Vanderbilt.scRNAseq_Fibroblasts, Epidish_BSeqSC.Vanderbilt.scRNAseq_Fibroblasts, CBSX_BSeqSC.Vanderbilt.scRNAseq_Fibroblasts, DWLS_BSeqSC.Vanderbilt.scRNAseq_Fibroblasts"
#> Keeping only feature: DeconRNASeq_BSeqSC.Vanderbilt.scRNAseq_Fibroblasts
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_CBSX.Vanderbilt.scRNAseq_Fibroblasts, Epidish_CBSX.Vanderbilt.scRNAseq_Fibroblasts, CBSX_CBSX.Vanderbilt.scRNAseq_Fibroblasts, DWLS_CBSX.Vanderbilt.scRNAseq_Fibroblasts, DeconRNASeq_DWLS.Vanderbilt.scRNAseq_Fibroblasts, Epidish_DWLS.Vanderbilt.scRNAseq_Fibroblasts, CBSX_DWLS.Vanderbilt.scRNAseq_Fibroblasts, DWLS_DWLS.Vanderbilt.scRNAseq_Fibroblasts, DWLS_MOMF.Vanderbilt.scRNAseq_Fibroblasts, Bisque_Vanderbilt_Fibroblasts"
#> Keeping only feature: Epidish_DWLS.Vanderbilt.scRNAseq_Fibroblasts
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_MOMF.Vanderbilt.scRNAseq_Fibroblasts, CBSX_MOMF.Vanderbilt.scRNAseq_Fibroblasts, AutogeneS_Vanderbilt_Fibroblasts, BayesPrism_Vanderbilt_Fibroblasts, MuSic_Vanderbilt_Fibroblasts, SCDC_Vanderbilt_Fibroblasts"
#> Keeping only feature: DeconRNASeq_MOMF.Vanderbilt.scRNAseq_Fibroblasts
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_BSeqSC.Vanderbilt.scRNAseq_Mast.cells, CBSX_BSeqSC.Vanderbilt.scRNAseq_Mast.cells, DWLS_BSeqSC.Vanderbilt.scRNAseq_Mast.cells, Epidish_DWLS.Vanderbilt.scRNAseq_Mast.cells, DeconRNASeq_MOMF.Vanderbilt.scRNAseq_Mast.cells"
#> Keeping only feature: DeconRNASeq_MOMF.Vanderbilt.scRNAseq_Mast.cells
#> 
#> [1] "Highly correlated features (r>0.9): DWLS_CBSX.Vanderbilt.scRNAseq_Mast.cells, DeconRNASeq_DWLS.Vanderbilt.scRNAseq_Mast.cells, CBSX_DWLS.Vanderbilt.scRNAseq_Mast.cells, DWLS_DWLS.Vanderbilt.scRNAseq_Mast.cells, Epidish_MOMF.Vanderbilt.scRNAseq_Mast.cells, CBSX_MOMF.Vanderbilt.scRNAseq_Mast.cells, DWLS_MOMF.Vanderbilt.scRNAseq_Mast.cells, BayesPrism_Vanderbilt_Mast.cells, Bisque_Vanderbilt_Mast.cells, SCDC_Vanderbilt_Mast.cells"
#> Keeping only feature: DWLS_CBSX.Vanderbilt.scRNAseq_Mast.cells
#> 
#> [1] "Highly correlated features (r>0.9): Epidish_LM22_Mast.activated.cells, CBSX_LM22_Mast.activated.cells, DWLS_LM22_Mast.activated.cells"
#> Keeping only feature: Epidish_LM22_Mast.activated.cells
#> 
#> [1] "Highly correlated features (r>0.9): DeconRNASeq_CBSX.HNSCC.scRNAseq_CAF, Epidish_CBSX.HNSCC.scRNAseq_CAF, CBSX_CBSX.HNSCC.scRNAseq_CAF, DWLS_CBSX.HNSCC.scRNAseq_CAF, DeconRNASeq_CBSX.Melanoma.scRNAseq_CAF, Epidish_CBSX.Melanoma.scRNAseq_CAF, CBSX_CBSX.Melanoma.scRNAseq_CAF, DWLS_CBSX.Melanoma.scRNAseq_CAF"
#> Keeping only feature: DeconRNASeq_CBSX.HNSCC.scRNAseq_CAF
#> 
#> Warning: Deconvolution features with less than two columns for subgrouping (skipping)
#> Warning: Deconvolution features with less than two columns for subgrouping (skipping)
#> Warning: Deconvolution features with less than two columns for subgrouping (skipping)
#> Warning: Deconvolution features with less than two columns for subgrouping (skipping)
#> Warning: Deconvolution features with less than two columns for subgrouping (skipping)
#> Warning: Deconvolution features with less than two columns for subgrouping (skipping)
#> Warning: Deconvolution features with less than two columns for subgrouping (skipping)
#> Warning: Deconvolution features with less than two columns for subgrouping (skipping)
#> Warning: Deconvolution features with less than two columns for subgrouping (skipping)
#> Warning: Deconvolution features with less than two columns for subgrouping (skipping)
#> Warning: Deconvolution features with less than two columns for subgrouping (skipping)
#> Warning: Deconvolution features with less than two columns for subgrouping (skipping)
#> Warning: Deconvolution features with less than two columns for subgrouping (skipping)
#> Warning: Deconvolution features with less than two columns for subgrouping (skipping)
#> Warning: Deconvolution features with less than two columns for subgrouping (skipping)
#> Warning: Deconvolution features with less than two columns for subgrouping (skipping)
#> Deconvolution features subgroupped